Move _force_unicode and _force_utf8 to their own module
Having them in filters was causing issues importing them in souptest
